+++ src/map-interface/Settings.ts
