+++ src/map-interface/components/CloseableCard.tsx
