+++ src/map-interface/components/LithChip.tsx
