+++ src/map-interface/components/elevation-chart.tsx
