+++ src/map-interface/components/gdd/Article.tsx
