+++ src/map-interface/components/gdd/Journal.tsx
