+++ src/map-interface/components/icons/BedrockIcon.tsx
