+++ src/map-interface/components/icons/ColumnIcon.tsx
