+++ src/map-interface/components/icons/ElevationIcon.tsx
