+++ src/map-interface/components/icons/FossilIcon.tsx
