+++ src/map-interface/components/icons/LineIcon.tsx
