+++ src/map-interface/components/info-drawer/index.tsx
