+++ src/map-interface/components/settings-panel.ts
