+++ src/map-interface/map-page/map-view/index.ts
