+++ src/map-interface/reducers/index.ts
