+++ src/map-interface/reducers/menu.ts
